LogistiQ/requirements.txt
anthropic>=0.50.0
redis>=5.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
